refactor: Layout Component 폴더 구조 수정
