+++ app/(scrollable-layout)/search/page.tsx
@@ -1,8 +1,8 @@
 import { Suspense } from "react";
 
 import ScrollToTop from "@/components/scroll-to-top";
-import PageContainer from "@/components/page-container";
-import ScrollLockOnce from "@/components/page-only/scroll-lock-once";
+import PageContainer from "@/components/layout/page-container";
+import ScrollLockOnce from "@/components/layout/scroll-lock-once";
 import QueryResolver from "@/features/search/components/query-resolver";
 import Header from "./components/result-header";
 import Contents from "./components/result-contents";
