+++ app/layout.tsx
@@ -6,9 +6,8 @@ import { playfair, urbanist } from "@/config/fonts";
 
 import "@/globals.css";
 import Header from "@/components/header";
-import CheckDevelopment from "@/components/check-development";
 import StateProvider from "@/components/state-provider";
-import BodyWithScrollLock from "@/components/body-with-scroll-lock";
+import BodyWithScrollLock from "@/components/layout/body-with-scroll-lock";
 import ThemeProvider from "@/features/theme/components/theme-provider";
 
 // 웹 페이지의 기본 메타데이터이다.
@@ -47,7 +46,6 @@ export default function RootLayout({
           <ThemeProvider>
             <Header />
             {children}
-            <CheckDevelopment />
             <Toaster toastOptions={{ position: "bottom-right" }} />
           </ThemeProvider>
         </BodyWithScrollLock>
